Move to openFrameworks version 0.10.1
If using the xcode project, you have to change the xcode project settings so
that it can find the LeapMotion SDK header and library files. You may also
have to adjust the linker runtime flags so that it loads the LeapMotion dylib
correctly (or just copy libLeap.dylib to bin/SonOfGesterBallDebug.app/Contents/MacOS/).